requirements.txt
requests>=2.31.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
playwright>=1.40.0
websockets>=12.0
aiohttp>=3.9.0
flask>=3.0.0
qrcode>=7.4.0
pillow>=10.0.0
openai>=1.0
psycopg2-binary>=2.9.0

# === LLM-RL 하이브리드 분산 시스템 (Phase 1) ===
pyzmq>=25.0
msgpack>=1.0
google-generativeai>=0.8.0

# === Phase 2+3 (RL 분산 학습) ===
gymnasium>=0.29
numpy>=1.24.0
torch>=2.0
# stable-baselines3>=2.0  # SB3 단일머신 훈련 시 활성화 (분산은 자체 PPO 사용)
